requirements.txt
Flask==1.0.2
pymongo==3.7.2
python-dotenv==0.9.1
